+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # CHANGELOG  
- 
+  
+## v1.0.4  
+  
+- Added BlendedRandom() class which blends between two Random sources  
+  
 ## v1.0.3  
+  
 - Made public consts for plugin info  
 - Added BepInEx soft dependency to Lethal Config  
 - Fixed LethalConfigNicerizer making duplicate pages in Lethal Config  
- 
+  
 ## v1.0.2  
+  
 - Fixed StaticNetcodeLib being included in the build output  
   
 ## v1.0.1